python-example/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
deprecated==1.2.18
fastapi==0.115.12
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.14.0
idna==3.10
importlib-metadata==8.6.1
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
packaging==25.0
protobuf==5.29.4
pydantic==2.11.3
pydantic-core==2.33.1
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
wrapt==1.17.2
zipp==3.21.0
